+++ index.d.ts
@@ -7,6 +7,8 @@ import { Agent, ClientRequest, RequestOptions } from 'http';
 
 export declare interface ProxyAgentOptions {
 	resolveProxy(req: ClientRequest, opts: RequestOptions, url: string, callback: (proxy: string) => void): void;
+	defaultPort: number;
+	originalAgent?: Agent;
 }
 
 export declare class ProxyAgent extends Agent {