+++ tests/test-client/src/utils.ts
@@ -2,13 +2,14 @@ import * as http from 'http';
 import * as https from 'https';
 import * as fs from 'fs';
 import * as path from 'path';
+import * as assert from 'assert';
 
 export const ca = [
 	fs.readFileSync(path.join(__dirname, '../../test-https-server/ssl_cert.pem')),
 	fs.readFileSync(path.join(__dirname, '../../test-https-server/ssl_teapot_cert.pem')),
 ];
 
-export async function testRequest<C extends typeof https | typeof http>(client: C, options: C extends typeof https ? https.RequestOptions : http.RequestOptions) {
+export async function testRequest<C extends typeof https | typeof http>(client: C, options: C extends typeof https ? https.RequestOptions : http.RequestOptions, testOptions: { assertResult?: (result: any) => void; } = {}) {
 	return new Promise<void>((resolve, reject) => {
 		const req = client.request(options, res => {
 			if (!res.statusCode || res.statusCode < 200 || res.statusCode > 299) {
@@ -20,10 +21,16 @@ export async function testRequest<C extends typeof https | typeof http>(client: 
 				data += chunk;
 			});
 			res.on('end', () => {
-				if (data === 'OK!') {
+				try {
+					const result = JSON.parse(data);
+					assert.equal(result.status, 'OK!');
+					if (testOptions.assertResult) {
+						testOptions.assertResult(result);
+					}
 					resolve();
-				} else {
-					reject(new Error(`Unexpected response: ${data}`));
+				} catch (err) {
+					err.message = `${err.message}: ${data}`;
+					reject(err);
 				}
 			});
 		});