+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 Notable changes will be documented here.
 
+## [0.4.0]
+- Fall back to original agent when provided in options.
+- Add default port to options.
+
 ## [0.3.0]
 - Forward request and options to `resolveProxy`.